refactor: ページコンポーネントの不要なコードを削除
- PopulationLine.tsxの不要なimport文とコメントを削除しました。
- PrefCheckboxes.tsxのimport文をconstantsディレクトリに移動しました。
